fix: resolve all Kotlin compiler warnings (30+)
- Deprecated API: suppress MODE_BACKGROUND_ACTIVITY_START_ALLOWED, SOFT_INPUT_ADJUST_RESIZE,
  MediaRecorder(), Os.setgid/setuid/seteuid, invalidate(Rect), setBoundsInParent(Rect),
  EXTRA_SHORTCUT_NAME/INTENT, Icons.Filled.VolumeOff/ArrowForwardIos/Forward,
  rememberModalBottomSheetState
- Unnecessary safe call: suppress where nullable receiver is compiler-inferred non-null
- Redundant conversion: remove/annotate redundant .toLong()/.toFloat()/.toString()
- No cast needed: replace mapNotNull+as? with map
- Unchecked cast: add @Suppress annotation
- Unnecessary non-null assertion: add @Suppress where nullable guaranteed at call site
- Always-true instance check: remove redundant 'is' check
- @Suppress elevated to function level where when-branch annotation was ineffective

+++ app/src/main/java/com/slideindex/app/widget/WidgetCanvasLayout.kt
@@ -406,7 +406,7 @@ class WidgetCanvasLayout(context: Context) : ViewGroup(context) {
       val top = paddingTop + hoverCellY * step
       val right = left + item.spanX * step
       val bottom = top + item.spanY * step
-      dragPreviewRect.set(left.toFloat(), top.toFloat(), right.toFloat(), bottom.toFloat())
+      dragPreviewRect.set(left, top, right, bottom)
       canvas.drawRoundRect(dragPreviewRect, cornerRadius, cornerRadius, dragPreviewPaint)
     }
   }
